Allow editing of basic profile info
