+++ app/controllers/profile_picture_uploader_controller.rb
@@ -1,0 +1,13 @@
+class ProfilePictureUploaderController < ApplicationController
+  def image
+    set_rack_response ProfilePictureUploader.upload_response(:cache, request.env)
+  end
+
+  private
+
+  def set_rack_response((status, headers, body))
+    self.status = status
+    self.headers.merge!(headers)
+    self.response_body = body
+  end
+end
